backend/pyproject.toml
[project]
name = "fast-room-api"
version = "0.1.0"
description = "FastAPI WebSocket API for FastRoom"
authors = [{ name = "John Aziz", email = "johnaziz269@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi[standard-no-fastapi-cloud-cli]==0.116.1",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "redis==6.4.0",
  "orjson==3.11.1",
  "httpx==0.28.1",
  "python-jose==3.5.0",
  "bcrypt==4.3.0",
  "sqlalchemy[asyncio]>=2.0.42",
  "python-dotenv>=1.1.1",
  "asyncpg>=0.30.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "pytest-cov>=6.2.1",
  "aiosqlite>=0.20.0",
]
lint = [
  "ruff>=0.12.8",
]
dev = [
  "mypy>=1.17.1",
  "types-redis>=4.6.0.20241004",
  "types-orjson>=3.6.2",
  "types-python-jose>=3.5.0.20250531",
  "pre-commit>=4.2.0",
  "alembic>=1.16.4",
]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
  ".venv",
  "dist",
  "build",
  "migrations",
]

[tool.ruff.lint]
select = [
  "E", "F", "I", "UP"
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
pretty = true
exclude = [
  ".venv",
  "tests",
  "frontend",
]

[tool.pytest.ini_options]
addopts = "-ra --cov"
# Tests reside under src/tests (keep future compatibility if moved)
testpaths = ["src/tests", "tests"]
pythonpath = ['backend']
asyncio_mode = "auto"

[tool.coverage.report]
show_missing = true
